DiffuseVAE/requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
hydra-core==1.3.2
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.2
kiwisolver==1.4.5
lmdb==1.5.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
multidict==6.0.5
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
packaging==24.0
pathtools==0.1.2
Pillow==9.5.0
platformdirs==4.0.0
promise==2.3
protobuf==3.20.3
psutil==6.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.9
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.7.3
sentry-sdk==2.19.2
setproctitle==1.3.3
shortuuid==1.0.9
six==1.17.0
smmap==5.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.6.0
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
wandb==0.18.7
Werkzeug==2.2.3
yarl==1.9.4
zipp==3.15.0
